+++ MapboxSearch/discover/src/test/java/com/mapbox/search/discover/DiscoverAddressTest.kt
@@ -4,18 +4,18 @@ import com.mapbox.search.common.tests.ToStringVerifier
 import nl.jqno.equalsverifier.EqualsVerifier
 import org.junit.jupiter.api.Test
 
-internal class DiscoverApiAddressTest {
+internal class DiscoverAddressTest {
 
     @Test
     fun `equals() and hashCode() functions are correct`() {
-        EqualsVerifier.forClass(DiscoverApiAddress::class.java)
+        EqualsVerifier.forClass(DiscoverAddress::class.java)
             .verify()
     }
 
     @Test
     fun `toString() function is correct`() {
         ToStringVerifier(
-            clazz = DiscoverApiAddress::class,
+            clazz = DiscoverAddress::class,
             includeAllProperties = false
         ).verify()
     }