+++ MapboxSearch/discover/src/test/java/com/mapbox/search/discover/DiscoverQueryTest.kt
@@ -5,18 +5,18 @@ import nl.jqno.equalsverifier.EqualsVerifier
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 
-internal class DiscoverApiQueryTest {
+internal class DiscoverQueryTest {
 
     @Test
     fun `equals() and hashCode() functions are correct`() {
-        EqualsVerifier.forClass(DiscoverApiQuery.Category::class.java)
+        EqualsVerifier.forClass(DiscoverQuery.Category::class.java)
             .verify()
     }
 
     @Test
     fun `toString() function is correct`() {
         ToStringVerifier(
-            clazz = DiscoverApiQuery.Category::class,
+            clazz = DiscoverQuery.Category::class,
             includeAllProperties = false
         ).verify()
     }
@@ -24,7 +24,7 @@ internal class DiscoverApiQueryTest {
     @Test
     fun `Check Category instantiation`() {
         val testCanonicalName = "test-canonical-name"
-        val categoryQuery = DiscoverApiQuery.Category.create(testCanonicalName)
+        val categoryQuery = DiscoverQuery.Category.create(testCanonicalName)
         assertEquals(testCanonicalName, categoryQuery.canonicalName)
     }
 }